refactor: 패키지명을 MVC 패턴에 알맞게 변경
