+++ src/main/java/com/nextstep/racingcar/Main.java
@@ -1,11 +1,11 @@
 package com.nextstep.racingcar;
 
-import com.nextstep.racingcar.racingcar.Simulator;
+import com.nextstep.racingcar.controller.RacingcarGame;
 
 public class Main {
 
   public static void main(String[] args) {
-    Simulator simulator = new Simulator();
-    simulator.play();
+    RacingcarGame racingcarGame = new RacingcarGame();
+    racingcarGame.play();
   }
 }
