+++ src/main/java/com/nextstep/racingcar/domain/CarResult.java
@@ -1,4 +1,4 @@
-package com.nextstep.racingcar.racingcar;
+package com.nextstep.racingcar.domain;
 
 public enum CarResult {
     STOP, MOVE
